rename application home to spo
