+++ spo/apps.py
@@ -10,4 +10,4 @@ class HomeConfig(AppConfig):
     """
 
     default_auto_field = "django.db.models.BigAutoField"
-    name = "home"
+    name = "spo"
