+++ templates/publication_generation_failed.html
@@ -37,7 +37,7 @@
     </div>
 </section>
 
-<a href="{% url 'home:publication_upload' %}">
+<a href="{% url 'spo:publication_upload' %}">
     Recommencer une nouvelle génération de publication
 </a>
