+++ templates/publication_generation_success.html
@@ -4,7 +4,7 @@
 {% block page_title %}Génération terminée{% endblock page_title %}
 
 {% block meta %}
-<meta http-equiv='refresh' content='0;url={% url "home:publication_generation_ended" generation_id %}'>
+<meta http-equiv='refresh' content='0;url={% url "spo:publication_generation_ended" generation_id %}'>
 {% endblock meta %}
 
 {% block content %}
@@ -12,7 +12,7 @@
 <p>
     Un téléchargement a démarré dans votre navigateur,
     si ce n'est pas le cas vous pouvez
-    <a href="{% url "home:publication_generation_ended" generation_id %}">
+    <a href="{% url "spo:publication_generation_ended" generation_id %}">
         lancer le téléchargement en cliquant ici
     </a>.
 </p>
